pictureGen/requirements.txt
# FLASK 
waitress
flask 
redis

# for auto-restarting
watchdog 

# for larnin'
tensorflow
textgenrnn
imageai

# other stuff
opencv-python
pillow
imutils
requests
awscli
boto3

